requirements.txt
python-dotenv
aiohttp
fastapi[all]
uvicorn
requests
daily-python
pillow
sounddevice
numpy